pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privacy-rag-onprem"
version = "0.2.0"
description = "Privacy-Preserving RAG for Clinical Data Systems — On-Premises Reference Implementation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "DanielMartin Arogyasami" },
]
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "slowapi>=0.1.9",

    # Document parsing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "defusedxml>=0.7.1",

    # Embeddings & ML
    "sentence-transformers>=3.0.0",
    "torch>=2.4.0",
    "transformers>=4.44.0",

    # Vector store
    "qdrant-client>=1.11.0",

    # Sparse retrieval
    "rank-bm25>=0.2.2",

    # PHI Detection
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "xxhash>=3.5.0",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "biopython>=1.83",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.6.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src", "config", "prompts"]
